Prototype support for BASIC authentication via server-feature interface.
